+++ app/scripts/SearchField.js
@@ -1,6 +1,9 @@
 import { bisector } from 'd3-array';
-import { format } from 'd3-format';
-import { absToChr } from './utils';
+import {
+  absToChr,
+  chrPosToPositionString,
+  convertNumberNotation,
+} from './utils';
 
 class SearchField {
   constructor(chromInfo) {
@@ -23,36 +26,7 @@ class SearchField {
     const y1 = absToChr(yScale.domain()[0], this.chromInfo);
     const y2 = absToChr(yScale.domain()[1], this.chromInfo);
 
-    let positionString = null;
-    const stringFormat = format(',d');
-
-    if (x1[0] !== x2[0]) {
-      // different chromosomes
-
-      positionString = `${x1[0]}:${stringFormat(Math.floor(x1[1]))}-${
-        x2[0]
-      }:${stringFormat(Math.ceil(x2[1]))}`;
-    } else {
-      // same chromosome
-
-      positionString = `${x1[0]}:${stringFormat(
-        Math.floor(x1[1])
-      )}-${stringFormat(Math.ceil(x2[1]))}`;
-    }
-
-    if (twoD) {
-      if (y1[0] !== y2[0]) {
-        // different chromosomes
-        positionString += ` & ${y1[0]}:${stringFormat(Math.floor(y1[1]))}-${
-          y2[0]
-        }:${stringFormat(Math.ceil(y2[1]))}`;
-      } else {
-        // same chromosome
-        positionString += ` & ${y1[0]}:${stringFormat(
-          Math.floor(y1[1])
-        )}-${stringFormat(Math.ceil(y2[1]))}`;
-      }
-    }
+    let positionString = chrPosToPositionString(x1, x2, y1, y2, twoD);
 
     if (x1[2] <= 0 || x2[2] > 0 || (twoD && (y1[2] <= 0 || y2[2] > 0))) {
       // did any of the coordinates exceed the genome boundaries
@@ -67,59 +41,6 @@ class SearchField {
     return positionString;
   }
 
-  convertNumberNotation(numStr) {
-    // Convert K or M notations
-    // e.g. "1.5M" to "1500000"
-    // or "0.05M" to "50000"
-    // or even "00.05M" or "00.050M" to "50000"
-    let newNumStr = numStr;
-
-    if (
-      !newNumStr.includes('M', newNumStr.length - 1) &&
-      !newNumStr.includes('K', newNumStr.length - 1)
-    ) {
-      // Nothing to convert
-      return newNumStr;
-    }
-
-    let numZerosToAdd = 0;
-    let decPointPosFromEnd = 0;
-
-    // Handle 'M' or 'N' notations
-    if (newNumStr.includes('M', newNumStr.length - 1)) {
-      numZerosToAdd = 6;
-      newNumStr = newNumStr.replace('M', '');
-    } else {
-      numZerosToAdd = 3;
-      newNumStr = newNumStr.replace('K', '');
-    }
-
-    if (Number.isNaN(+newNumStr)) {
-      // Without 'K' or 'M' notation, the string should be converted to a valid number.
-      return numStr;
-    }
-
-    // Drop the needless characters for the simplicity (e.g., "00.5" to "0.5" or "1,000" to "1000").
-    newNumStr = (+newNumStr).toString();
-
-    // Handle a decimal point
-    if (newNumStr.includes('.')) {
-      decPointPosFromEnd = newNumStr.length - 1 - newNumStr.indexOf('.');
-      newNumStr = (+newNumStr.replace('.', '')).toString();
-    }
-
-    const totalZerosToAdd = numZerosToAdd - decPointPosFromEnd;
-    if (totalZerosToAdd < 0) {
-      // The value is smaller than 1 (e.g. "0.00005K")
-      return numStr;
-    }
-
-    // Finally, add zeros at the end.
-    newNumStr += '0'.repeat(totalZerosToAdd);
-
-    return newNumStr;
-  }
-
   parsePosition(positionText, prevChr = null) {
     // Parse chr:position strings...
     // i.e. chr1:1000
@@ -130,7 +51,7 @@ class SearchField {
 
     if (positionParts.length > 1) {
       chr = positionParts[0];
-      pos = +this.convertNumberNotation(positionParts[1].replace(/,/g, '')); // chromosome specified
+      pos = +convertNumberNotation(positionParts[1].replace(/,/g, '')); // chromosome specified
     } else if (positionParts[0] in this.chromInfo.chrPositions) {
       // is this an entire chromosome
       chr = positionParts[0];
@@ -143,7 +64,7 @@ class SearchField {
       }
     } else {
       // no it's just a position without a chromosome
-      pos = +this.convertNumberNotation(positionParts[0].replace(/,/g, '')); // no chromosome specified
+      pos = +convertNumberNotation(positionParts[0].replace(/,/g, '')); // no chromosome specified
       chr = null;
 
       if (prevChr) chr = prevChr;
@@ -238,7 +159,7 @@ class SearchField {
       // it's length as the range
       range = [
         chromPosition,
-        chromPosition + +this.chromInfo.chromLengths[parts[0]]
+        chromPosition + +this.chromInfo.chromLengths[parts[0]],
       ];
     } else {
       // e.g. ("chr1:540340")
@@ -269,7 +190,7 @@ class SearchField {
     if (parts.length === 0) {
       return [
         [0, 0],
-        [0, 0]
+        [0, 0],
       ];
     }
 
@@ -277,14 +198,14 @@ class SearchField {
       const sparts = parts[0].split(',');
       return [
         [+sparts[0], +sparts[1]],
-        [0, 0]
+        [0, 0],
       ];
     }
     const sparts0 = parts[0].split(',');
     const sparts1 = parts[1].split(',');
     return [
       [+sparts0[0], +sparts0[1]],
-      [+sparts1[0], +sparts1[1]]
+      [+sparts1[0], +sparts1[1]],
     ];
   }
 
@@ -301,7 +222,7 @@ class SearchField {
     // or the distance after the last chromosome of the given
     let offset = [
       [0, 0],
-      [0, 0]
+      [0, 0],
     ];
     if (offsetRe) {
       text = text.replace(offsetRe[0], '');